@@ -3,20 +3,14 @@ import text from './example.js';
 import { MuiThemeProvider } from 'material-ui/styles';
 import AvPlayArrow from 'material-ui/svg-icons/av/play-arrow';
 import AvPause from 'material-ui/svg-icons/av/pause';
-import injectTapEventPlugin from 'react-tap-event-plugin';
 import Player from './Player';
 import Editor from './Editor';
-import soundfonts from 'soundfonts';
-import guitar from 'soundfonts/instruments/Acoustic Guitar (nylon)/mp3.json';
-// import ogg from 'soundfonts/instruments/Acoustic Guitar (nylon)/ogg.json';
 
 import {
   Drawer,
   AppBar,
   IconButton,
 } from 'material-ui';
-
-injectTapEventPlugin();
 
 const GithubLink = ({
   touch,
@@ -27,19 +21,19 @@ const GithubLink = ({
 }) =>
   <IconButton { ...{ touch, target, iconStyle } }
     iconClassName="muidocs-icon-custom-github"
-    href={'https://github.com/' + user + '/' + name}
+    href={`https://github.com/${user}/${name}`}
     tooltip="Github"
     tooltipPosition="bottom-left"
   />;
 
-class App extends Component {
+export default class App extends Component {
 
   state = {
     open: false,
     playing: false,
     text: text,
     tempo: 120,
-    context: new AudioContext(),
+    strings: [],
     ranges: [{
       anchor: { ch: 0, line: 0 },
       head: { ch: 0, line: 0 },
@@ -53,6 +47,7 @@ class App extends Component {
     this.setState({ playing: !this.state.playing });
   }
 
+  // Move the cursors along by one step.
   tick =() => {
     var ranges = this.state.ranges.map(({ anchor: { ch, line }}) => {
       var next = text.split('\n')[line].slice(ch);
@@ -60,15 +55,15 @@ class App extends Component {
         ch: ch + 1,
         line,
       }
-      if (next[0] === '|') {
-        if (next[1] === '-') {
-          anchor.ch += 4;
-        } else if (!next[1]) {
+      // if (next[0] === '|') {
+      //   if (next[1] === '-') {
+      //     anchor.ch += 4;
+      //   } else if (!next[1]) {
 
-          anchor.ch = text.split('\n')[line + 10].indexOf('|---') + 5;
-          anchor.line += 10;
-        }
-      }
+      //     anchor.ch = text.split('\n')[line + 10].indexOf('|---') + 5;
+      //     anchor.line += 10;
+      //   }
+      // }
       return {
         anchor, head: {
           line: anchor.line,
@@ -79,9 +74,16 @@ class App extends Component {
     this.setState({ ranges })
   };
 
-  cursorActivity = instance => this.setState({
-    ranges: instance.listSelections()
-  });
+  cursorActivity = instance => {
+    console.log('cursorActivity', { instance })
+    this.setState({
+      strings: instance.listSelections().map(({
+        anchor: { line, ch },
+        head,
+      }) => instance.getLine(line).slice(0, ch)),
+      ranges: instance.listSelections()
+    });
+  }
 
   update = text => this.setState({ text });
 
@@ -109,71 +111,118 @@ class App extends Component {
       .then(instrument => this.setState({ instrument }));
   }
 
-  render = () =>
-    <MuiThemeProvider>
-      <div style={this.props.style}>
+  skip = key => () => {
+    var ranges = this.state.ranges.map(({ anchor: { ch, line }}) => {
+      var next = text.split('\n')[line].slice(ch);
+      var anchor = {
+        ch: ch + 1,
+        line,
+      }
+      // if (next[0] === '|') {
+      //   if (next[1] === '-') {
+      //     anchor.ch += 4;
+      //   } else if (!next[1]) {
+
+      //     anchor.ch = text.split('\n')[line + 10].indexOf('|---') + 5;
+      //     anchor.line += 10;
+      //   }
+      // }
+      return {
+        anchor, head: {
+          line: anchor.line,
+          ch: anchor.ch - 1,
+        },
+      }
+    });
+    this.setState(({ ranges }) => ({ ranges:  }))
+  }
+
+  render = () => {
+    const {
+      jump,
+      toggle,
+      update,
+      cursorActivity,
+      state: {
+        playing,
+        open,
+        ranges,
+        strings,
+        text,
+      },
+      props: {
+        style,
+      },
+    } = this;
+
+    const options = {
+      lineNumbers: true,
+      viewportMargin: Infinity,
+      firstLineNumber: 0,
+      fixedGutter: false,
+    }
+
+    const iconElementRight = <GithubLink
+      touch
+      target="_blank"
+      user='ryanwhite04'
+      name='react-songs'
+      iconStyle={{
+        color: 'white',
+      }}
+    />
+
+    const iconElementLeft = <IconButton
+      touch
+      tooltip={playing ? 'Stop' : 'Play'}
+      tooltipPosition="bottom-right"
+      >
+      {playing ? <AvPause /> : <AvPlayArrow />}
+    </IconButton>
+  
+    return <MuiThemeProvider>
+      <div style={style}>
         <AppBar title="React Songs"
-          iconElementRight={
-            <GithubLink
-              touch
-              target="_blank"
-              user='ryanwhite04'
-              name='react-songs'
-              iconStyle={{
-                color: 'white',
-              }}
-            />
-          }
-          iconElementLeft={
-            <IconButton
-              touch
-              tooltip={this.state.playing ? 'Stop' : 'Play'}
-              tooltipPosition="bottom-right"
-              >
-              {this.state.playing ? <AvPause /> : <AvPlayArrow />}
-            </IconButton>
-          }
-          onLeftIconButtonTouchTap={this.toggle}
+
+          iconElementRight={iconElementRight}
+          iconElementLeft={iconElementLeft}
+          onLeftIconButtonClick={toggle}
           zDepth={2}
         />
         <Drawer
-          docked={false} open={this.state.open}
-          onRequestChange={open => this.setState({open})} >
+          docked={false} open={open}
+          onRequestChange={open => this.setState({ open })} >
         </Drawer>
         <Editor
           implementation='ace editor'
-          update={this.update}
-          cursorActivity={this.cursorActivity}
-          ranges={this.state.ranges}
-          options={{
-            lineNumbers: true,
-            viewportMargin: Infinity,
-            firstLineNumber: 0,
-            fixedGutter: false,
-          }}
+          update={update}
+          cursorActivity={cursorActivity}
+          ranges={ranges}
+          options={options}
           >
-          {this.state.text}
+          {text}
         </Editor>
-        <Player
-          playing={this.state.playing}
-          context={this.state.context}
-          ranges={this.state.ranges}
-          text={this.state.text}
-        >
-          {/* {guitar.map((src, i) => <audio
-            key={i}
-            crossOrigin="anonymous"
-            src={src} >
-          </audio>)} */}
-
-          {guitar.map((src, i) => {
-            var audio = new Audio(src);
-            audio.crossOrigin = 'anonymous';
-            return audio;
-          })}
-        </Player>
+        {strings.map((string, key) => <Player jump={jump(key)} key={key}>{string}</Player>)}
       </div>
     </MuiThemeProvider>
+  }
 }
 
-export default App;
+// let Player = <Player
+//   playing={playing}
+//   context={context}
+//   ranges={ranges}
+//   text={text}
+//   >
+//   {/* {guitar.map((src, i) => <audio
+//     key={i}
+//     crossOrigin="anonymous"
+//     src={src} >
+//   </audio>)} */}
+
+//   {guitar.map((src, i) => {
+//     var audio = new Audio(src);
+//     audio.crossOrigin = 'anonymous';
+//     return audio;
+//   })}
+// </Player>
